EC2_NFLsearching-master/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
fastapi==0.61.2
h11==0.14.0
idna==3.4
pydantic==2.3.0
pydantic_core==2.6.3
sniffio==1.3.0
typing_extensions==4.7.1
uvicorn==0.23.2
python-dotenv
mysql-connector
sqlalchemy
graphene
graphene-pydantic
orator
starlette==0.13.6
